Test eligibility calculator directly.
